refactor: rename Multi -> CrossField and move *With validators to String
The four startsWith/endsWith/doesntStartWith/doesntEndWith validators are
purely string operations; their cross-field framing was an accident of
their varargs signature.

Refs #23
