+++ src/lib/directives/cross-field/nguard-confirmed.directive.ts
@@ -1,24 +1,24 @@
 import { Directive, input } from '@angular/core';
 import { AbstractControl, NG_VALIDATORS, ValidationErrors, Validator } from '@angular/forms';
-import { MultiValidators } from '../../validators/multi.validators';
+import { CrossFieldValidators } from '../../validators/cross-field.validators';
 
 @Directive({
     providers: [
         {
             multi: true,
             provide: NG_VALIDATORS,
-            useExisting: NguardGreaterThanDirective,
+            useExisting: NguardConfirmedDirective,
         },
     ],
-    selector: '[nguardGreaterThan]',
+    selector: '[nguardConfirmed]',
     standalone: true,
 })
-export class NguardGreaterThanDirective implements Validator {
-    public readonly fieldKey = input.required<string>({ alias: 'nguardGreaterThan' });
+export class NguardConfirmedDirective implements Validator {
+    public readonly fieldKey = input.required<string>({ alias: 'nguardConfirmed' });
 
     constructor() {}
 
     public validate(control: AbstractControl<any, any>): ValidationErrors | null {
-        return MultiValidators.greaterThan(this.fieldKey())(control);
+        return CrossFieldValidators.confirmed(this.fieldKey())(control);
     }
 }
