+++ src/lib/directives/cross-field/nguard-greater-than-or-equal.directive.ts
@@ -1,6 +1,6 @@
 import { Directive, input } from '@angular/core';
 import { AbstractControl, NG_VALIDATORS, ValidationErrors, Validator } from '@angular/forms';
-import { MultiValidators } from '../../validators/multi.validators';
+import { CrossFieldValidators } from '../../validators/cross-field.validators';
 
 @Directive({
     providers: [
@@ -19,6 +19,6 @@ export class NguardGreaterThanOrEqualDirective implements Validator {
     constructor() {}
 
     public validate(control: AbstractControl<any, any>): ValidationErrors | null {
-        return MultiValidators.greaterThanOrEqual(this.fieldKey())(control);
+        return CrossFieldValidators.greaterThanOrEqual(this.fieldKey())(control);
     }
 }
