+++ src/lib/directives/cross-field/nguard-greater-than.directive.ts
@@ -1,24 +1,24 @@
 import { Directive, input } from '@angular/core';
 import { AbstractControl, NG_VALIDATORS, ValidationErrors, Validator } from '@angular/forms';
-import { MultiValidators } from '../../validators/multi.validators';
+import { CrossFieldValidators } from '../../validators/cross-field.validators';
 
 @Directive({
     providers: [
         {
             multi: true,
             provide: NG_VALIDATORS,
-            useExisting: NguardConfirmedDirective,
+            useExisting: NguardGreaterThanDirective,
         },
     ],
-    selector: '[nguardConfirmed]',
+    selector: '[nguardGreaterThan]',
     standalone: true,
 })
-export class NguardConfirmedDirective implements Validator {
-    public readonly fieldKey = input.required<string>({ alias: 'nguardConfirmed' });
+export class NguardGreaterThanDirective implements Validator {
+    public readonly fieldKey = input.required<string>({ alias: 'nguardGreaterThan' });
 
     constructor() {}
 
     public validate(control: AbstractControl<any, any>): ValidationErrors | null {
-        return MultiValidators.confirmed(this.fieldKey())(control);
+        return CrossFieldValidators.greaterThan(this.fieldKey())(control);
     }
 }
