+++ src/lib/directives/string/nguard-doesnt-end-with.directive.spec.ts
