+++ src/lib/directives/string/nguard-doesnt-start-with.directive.spec.ts
