+++ src/lib/directives/string/nguard-ends-with.directive.spec.ts
