+++ src/lib/directives/string/nguard-ends-with.directive.ts
@@ -5,27 +5,27 @@ import { AbstractControl, NG_VALIDATORS, ValidationErrors, Validator } from '@an
 import { primitive } from '../../utils/validators.utils';
 
 // Validators
-import { MultiValidators } from '../../validators/multi.validators';
+import { StringValidators } from '../../validators/string.validators';
 
 @Directive({
     providers: [
         {
             multi: true,
             provide: NG_VALIDATORS,
-            useExisting: NguardDoesntEndWithDirective,
+            useExisting: NguardEndsWithDirective,
         },
     ],
-    selector: '[nguardDoesntEndWith]',
+    selector: '[nguardEndsWith]',
     standalone: true,
 })
-export class NguardDoesntEndWithDirective implements Validator {
-    public readonly values = input.required<primitive | primitive[]>({ alias: 'nguardDoesntEndWith' });
+export class NguardEndsWithDirective implements Validator {
+    public readonly values = input.required<primitive | primitive[]>({ alias: 'nguardEndsWith' });
 
     constructor() {}
 
     public validate(control: AbstractControl<any, any>): ValidationErrors | null {
         const raw = this.values();
         const values = Array.isArray(raw) ? raw : [raw];
-        return MultiValidators.doesntEndWith(...values)(control);
+        return StringValidators.endsWith(...values)(control);
     }
 }
