+++ src/lib/directives/string/nguard-starts-with.directive.ts
@@ -5,27 +5,27 @@ import { AbstractControl, NG_VALIDATORS, ValidationErrors, Validator } from '@an
 import { primitive } from '../../utils/validators.utils';
 
 // Validators
-import { MultiValidators } from '../../validators/multi.validators';
+import { StringValidators } from '../../validators/string.validators';
 
 @Directive({
     providers: [
         {
             multi: true,
             provide: NG_VALIDATORS,
-            useExisting: NguardEndsWithDirective,
+            useExisting: NguardStartsWithDirective,
         },
     ],
-    selector: '[nguardEndsWith]',
+    selector: '[nguardStartsWith]',
     standalone: true,
 })
-export class NguardEndsWithDirective implements Validator {
-    public readonly values = input.required<primitive | primitive[]>({ alias: 'nguardEndsWith' });
+export class NguardStartsWithDirective implements Validator {
+    public readonly values = input.required<primitive | primitive[]>({ alias: 'nguardStartsWith' });
 
     constructor() {}
 
     public validate(control: AbstractControl<any, any>): ValidationErrors | null {
         const raw = this.values();
         const values = Array.isArray(raw) ? raw : [raw];
-        return MultiValidators.endsWith(...values)(control);
+        return StringValidators.startsWith(...values)(control);
     }
 }
