+++ src/lib/validators/nguard.validators.ts
@@ -1,9 +1,9 @@
-import { MultiValidators } from './multi.validators';
+import { CrossFieldValidators } from './cross-field.validators';
 import { NumberValidators } from './number.validators';
 import { StringValidators } from './string.validators';
 
 export namespace NguardValidators {
-    export const Multi = MultiValidators;
+    export const CrossField = CrossFieldValidators;
     export const Number = NumberValidators;
     export const String = StringValidators;
 }
